+++ Sources/SpotifyExampleContent/ExampleAudioAnalysis.swift
@@ -1,0 +1,11 @@
+import Foundation
+import SpotifyWebAPI
+
+public extension AudioAnalysis {
+    
+    /// Sample data for testing purposes.
+    static let anyColourYouLike = Bundle.module.decodeJson(
+        forResource: "Any Colour You Like - AudioAnalysis", type: Self.self
+    )!
+
+}