+++ Sources/SpotifyExampleContent/ExampleAudioFeatures.swift
@@ -1,0 +1,11 @@
+import Foundation
+import SpotifyWebAPI
+
+public extension AudioFeatures {
+    
+    /// Sample data for testing purposes.
+    static let fearless = Bundle.module.decodeJson(
+        forResource: "Fearless - AudioFeatures", type: Self.self
+    )!
+
+}