+++ Tests/SpotifyWebAPIMainTests/Coding Tests/CodingPlaybackRequestTests.swift
@@ -10,12 +10,6 @@ final class CodingPlaybackRequestTests: XCTestCase {
         ("testCodingPlaybackReqest", testCodingPlaybackReqest)
     ]
 
-    // override class func setUp() {
-    //     SpotifyDecodingError.dataDumpfolder = URL(fileURLWithPath:
-    //         "/Users/pschorn/Desktop/"
-    //     )
-    // }
-    
     func testCodingPlaybackReqest() throws {
         
         do {